+++ rust/operator-binary/src/security/opa.rs
@@ -54,7 +54,7 @@ impl HbaseOpaConfig {
     ) -> Result<Self> {
         let authorization_connection_string = authorization_config
             .opa
-            .full_document_url_from_config_map(client, hbase, Some("allow"), OpaApiVersion::V1)
+            .full_document_url_from_config_map(client, hbase, Some("allow"), &OpaApiVersion::V1)
             .await
             .context(ConstructOpaEndpointForAuthorizerSnafu)?;